factor out intro components
